+++ src/main/java/item/Item.java
@@ -7,12 +7,12 @@ public class Item {
     private int quantity;
     private final String uom;
     private final String category;
-    private int buyPrice;
-    private int sellPrice;
+    private float buyPrice;
+    private float sellPrice;
     private boolean isOOS;
 
 
-    public Item(String name, int quantity, String uom, String category, int buyPrice, int sellPrice) {
+    public Item(String name, int quantity, String uom, String category, float buyPrice, float sellPrice) {
         this.itemName = name;
         this.quantity = quantity;
         this.uom = uom;
@@ -53,10 +53,10 @@ public class Item {
         this.quantity = newQuantity;
     }
 
-    public int getBuyPrice() {
+    public float getBuyPrice() {
         return buyPrice;
     }
-    public int getSellPrice() {
+    public float getSellPrice() {
         return sellPrice;
     }
